requirements.txt
anyio==4.10.0
APScheduler==3.10.4
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
lxml==6.0.1
python-dotenv==1.1.1
python-telegram-bot==20.8
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
